Splits resolvers into multiple files
